requirements.txt
# Core AI Dependencies
openai>=1.0.0
langchain>=0.1.0
requests>=2.31.0
ollama>=0.2.0

# FastAPI & Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Memory & File Processing
pandas>=2.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.1.0

# Voice & Speech (Optional)
SpeechRecognition>=3.10.0
pyttsx3>=2.90
gTTS>=2.4.0
pyaudio>=0.2.11

# Image Processing (Optional)
Pillow>=10.0.0
pytesseract>=0.3.10

# Video Processing (Optional) 
moviepy>=1.0.3
whisper>=1.0.0

# Translation
deep-translator>=1.11.0

# Utilities
python-dotenv>=1.0.0
colorama>=0.4.6
pathlib>=1.0.1
